fix narrowing conversion in memory_manager_ga.h
Former-commit-id: edf7bcd633d8bd55d9a12d53d9978ff4398010e5
Former-commit-id: 3125b7d227a50c08f91df040640e08e770279d2c
Former-commit-id: f5a8738b23b3a107aa1d636abe1890c986956ed7
Former-commit-id: fe21e974fc21eb8f2af129e1f63965d9456e81bc [formerly 9321713a54d568abed631b59335270bb06149028]
Former-commit-id: 6abbd5c1f05abaa43a5ba07c4f436ff0d540c540

+++ tammx/memory_manager_ga.h
@@ -140,7 +140,7 @@ class MemoryManagerGA : public MemoryManager {
             allocation_status_ == AllocationStatus::attached);
     Proc proc{pg_.rank()};
     int64_t nels{1};
-    int iproc{proc.value()};
+    int64_t iproc{proc.value()};
     int64_t ioffset{map_[proc.value()] + off.value()};
     int64_t lo = ioffset, hi = ioffset + nels-1, ld = -1;
     void* buf;
@@ -153,7 +153,7 @@ class MemoryManagerGA : public MemoryManager {
             allocation_status_ == AllocationStatus::attached);
     Proc proc{pg_.rank()};
     int64_t nels{1};
-    int iproc{proc.value()};
+    int64_t iproc{proc.value()};
     int64_t ioffset{map_[proc.value()] + off.value()};
     int64_t lo = ioffset, hi = ioffset + nels-1, ld = -1;
     void* buf;
@@ -164,7 +164,7 @@ class MemoryManagerGA : public MemoryManager {
   void get(Proc proc, Offset off, Size nelements, void* buf) {
     EXPECTS(allocation_status_ == AllocationStatus::created ||
             allocation_status_ == AllocationStatus::attached);
-    int iproc{proc.value()};
+    int64_t iproc{proc.value()};
     int64_t ioffset{map_[proc.value()] + off.value()};
     int64_t lo = ioffset, hi = ioffset + nelements.value()-1, ld = -1;
     NGA_Get64(ga_, &lo, &hi, buf, &ld);
@@ -173,7 +173,7 @@ class MemoryManagerGA : public MemoryManager {
   void put(Proc proc, Offset off, Size nelements, const void* buf) {
     EXPECTS(allocation_status_ == AllocationStatus::created ||
             allocation_status_ == AllocationStatus::attached);
-    int iproc{proc.value()};
+    int64_t iproc{proc.value()};
     int64_t ioffset{map_[proc.value()] + off.value()};
     int64_t lo = ioffset, hi = ioffset + nelements.value()-1, ld = -1;
     NGA_Put64(ga_, &lo, &hi, const_cast<void*>(buf), &ld);
@@ -182,7 +182,7 @@ class MemoryManagerGA : public MemoryManager {
   void add(Proc proc, Offset off, Size nelements, const void* buf) {
     EXPECTS(allocation_status_ == AllocationStatus::created ||
             allocation_status_ == AllocationStatus::attached);
-    int iproc{proc.value()};
+    int64_t iproc{proc.value()};
     int64_t ioffset{map_[proc.value()] + off.value()};
     int64_t lo = ioffset, hi = ioffset + nelements.value()-1, ld = -1;
     void *alpha;
